Moved settings buttons into a different file
